requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.0
tzdata==2023.3
wcwidth==0.2.12
